Move time-related type to specific package
