+++ modules/core/src/main/scala/tofu/package.scala
@@ -11,4 +11,8 @@ package object tofu extends KernelTypes {
   type IorC[C[_], E, A] = Ior[C[E], C[A]]
 
   type Start[F[_]] = Fibers[F, Id, Fiber[F, *]]
+
+  type Timeout[F[_]] = tofu.time.Timeout[F]
+
+  val Timeout: tofu.time.Timeout.type = tofu.time.Timeout
 }
