+++ modules/kernel/src/main/scala/tofu/syntax/time.scala
@@ -1,11 +1,12 @@
 package tofu
 package syntax
+import tofu.time.{Clock, Sleep}
 
 import java.time.{Clock => _, _}
 import java.util.concurrent.TimeUnit
 
 import cats.{Apply, Functor}
-import tofu.common.{TimeData, TimeZone}
+import tofu.time.{TimeData, TimeZone}
 import tofu.syntax.monadic._
 import scala.concurrent.duration.FiniteDuration
